Updates on f32 support.

+++ bigtsne/_tsne.pxd
@@ -15,10 +15,10 @@ ctypedef fused sparse_index_type:
     np.int64_t
 
 
-cpdef double[:, ::1] compute_gaussian_perplexity(
-    double[:, :] distances,
-    double[:] desired_perplexities,
-    double perplexity_tol=*,
+cpdef float[:, ::1] compute_gaussian_perplexity(
+    float[:, :] distances,
+    float[:] desired_perplexities,
+    float perplexity_tol=*,
     Py_ssize_t max_iter=*,
     Py_ssize_t num_threads=*,
 )
@@ -26,76 +26,76 @@ cpdef double[:, ::1] compute_gaussian_perplexity(
 cpdef tuple estimate_positive_gradient_nn(
     sparse_index_type[:] indices,
     sparse_index_type[:] indptr,
-    double[:] P_data,
-    double[:, ::1] embedding,
-    double[:, ::1] reference_embedding,
-    double[:, ::1] gradient,
-    double dof=*,
+    float[:] P_data,
+    float[:, ::1] embedding,
+    float[:, ::1] reference_embedding,
+    float[:, ::1] gradient,
+    float dof=*,
     Py_ssize_t num_threads=*,
     bint should_eval_error=*,
 )
 
-cpdef double estimate_negative_gradient_bh(
+cpdef float estimate_negative_gradient_bh(
     QuadTree tree,
-    double[:, ::1] embedding,
-    double[:, ::1] gradient,
-    double theta=*,
-    double dof=*,
+    float[:, ::1] embedding,
+    float[:, ::1] gradient,
+    float theta=*,
+    float dof=*,
     Py_ssize_t num_threads=*,
     bint pairwise_normalization=*,
 )
 
-cpdef double estimate_negative_gradient_fft_1d(
-    double[::1] embedding,
-    double[::1] gradient,
+cpdef float estimate_negative_gradient_fft_1d(
+    float[::1] embedding,
+    float[::1] gradient,
     Py_ssize_t n_interpolation_points=*,
     Py_ssize_t min_num_intervals=*,
-    double ints_in_interval=*,
-    double dof=*,
+    float ints_in_interval=*,
+    float dof=*,
 )
 
 cpdef tuple prepare_negative_gradient_fft_interpolation_grid_1d(
-    double[::1] reference_embedding,
+    float[::1] reference_embedding,
     Py_ssize_t n_interpolation_points=*,
     Py_ssize_t min_num_intervals=*,
-    double ints_in_interval=*,
-    double dof=*,
-    double padding=*,
+    float ints_in_interval=*,
+    float dof=*,
+    float padding=*,
 )
 
-cpdef double estimate_negative_gradient_fft_1d_with_grid(
-    double[::1] embedding,
-    double[::1] gradient,
-    double[:, ::1] y_tilde_values,
-    double[::1] box_lower_bounds,
+cpdef float estimate_negative_gradient_fft_1d_with_grid(
+    float[::1] embedding,
+    float[::1] gradient,
+    float[:, ::1] y_tilde_values,
+    float[::1] box_lower_bounds,
     Py_ssize_t n_interpolation_points,
-    double dof,
+    float dof,
 )
 
-cpdef double estimate_negative_gradient_fft_2d(
-    double[:, ::1] embedding,
-    double[:, ::1] gradient,
+cpdef float estimate_negative_gradient_fft_2d(
+    float[:, ::1] embedding,
+    float[:, ::1] gradient,
     Py_ssize_t n_interpolation_points=*,
     Py_ssize_t min_num_intervals=*,
-    double ints_in_interval=*,
-    double dof=*,
+    float ints_in_interval=*,
+    float dof=*,
 )
 
 cpdef tuple prepare_negative_gradient_fft_interpolation_grid_2d(
-    double[:, ::1] reference_embedding,
+    float[:, ::1] reference_embedding,
     Py_ssize_t n_interpolation_points=*,
     Py_ssize_t min_num_intervals=*,
-    double ints_in_interval=*,
-    double dof=*,
-    double padding=*,
+    float ints_in_interval=*,
+    float dof=*,
+    float padding=*,
 )
 
-cpdef double estimate_negative_gradient_fft_2d_with_grid(
-    double[:, ::1] embedding,
-    double[:, ::1] gradient,
-    double[:, ::1] y_tilde_values,
-    double[::1] box_x_lower_bounds,
-    double[::1] box_y_lower_bounds,
+cpdef float estimate_negative_gradient_fft_2d_with_grid(
+    float[:, ::1] embedding,
+    float[:, ::1] gradient,
+    float[:, ::1] y_tilde_values,
+    float[::1] box_x_lower_bounds,
+    float[::1] box_y_lower_bounds,
     Py_ssize_t n_interpolation_points,
-    double dof,
+    float dof,
 )
